+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [6.4.0-pre.1] - 2026-01-??
 ### Fixed
 - Missing and existing deprecation warning messages.
+- Fixes on Custom HTML Elements.
 
 ## [6.3.0] - 2026-01-17
 **Stable Release**